little bit of goofy aah fixes and some basic functionality

+++ voice_analysis.py
@@ -1,24 +1,25 @@
-# import numpy as np
-# import soundfile as sf
-# from scipy.io.wavfile import read
-#
-# f = sf.SoundFile('295218314400235529.wav')
-# print(f.frames)
-# samprate, wavdata = read('295218314400235529.wav')
+import datetime
 
 from pydub import AudioSegment
-import glob, os, statistics, math
+import glob, os, statistics, math, asyncio
+import config
 
 
-def get_dbfs():
+def get_dbfs(guild_id):
     list = []
     cnt = 1
-    os.chdir('C:/Users/mosina/PycharmProjects/discord-voice-normalization-bot/recordings')
+    guild_recording_path = f'{config.recording_path}/{guild_id}'
+    os.chdir(guild_recording_path)
     for file in glob.glob('*.mp3'):
-        audio = AudioSegment.from_file(file).dBFS
-        list.append((file[0:len(file) - 4], audio))
-        cnt += 1
-        os.remove(file)
+        time_distinction = datetime.datetime.now() - datetime.datetime.fromtimestamp(os.path.getmtime(file))
+        if time_distinction.days >= 2:
+            os.remove(file)
+        else:
+            audio = AudioSegment.from_file(file).dBFS
+            list.append((file[0:len(file) - 4], audio))
+            cnt += 1
+
+    print(guild_id)
 
     penile = i_hate_math(list)
 
@@ -27,14 +28,14 @@ def get_dbfs():
 
 def i_hate_math(ListDbfs):
     timedlist = [x[1] for x in ListDbfs]
-    print(timedlist)
     array_of_perc = []
+    print(timedlist)
     med = statistics.median(timedlist)
     perc_med = math.pow(10, med / 10)
-    for i in timedlist:
-        j = math.pow(10, i / 10) / perc_med * 100
+    for i in range(len(timedlist)):
+        j = math.pow(10, timedlist[i] / 10) / perc_med * 100
         if 0 <= j <= 200:
-            array_of_perc.append(round(200 - j))
+            array_of_perc.append(200 - j)
         if j > 200:
             array_of_perc.append('Ты слишком громкий, бро.')
         if j < 0:
@@ -42,6 +43,10 @@ def i_hate_math(ListDbfs):
 
     return array_of_perc
 
+
+def delete_audio(guild_id, user_id):
+    user_recording_path = f'{config.recording_path}/{guild_id}/{user_id[2:-1]}.mp3'
+    os.remove(user_recording_path)
 
 # audio = AudioSegment.from_file('295218314400235529.wav')
 # print(audio.dBFS)
